server: CheckMovingGround should accept float as it called with float arg

+++ game/server/player_command.cpp
@@ -90,7 +90,8 @@ void CPlayerMove::FinishCommand( CBasePlayer *player )
 // Input  : *player - 
 //			frametime - 
 //-----------------------------------------------------------------------------
-void CPlayerMove::CheckMovingGround( CBasePlayer *player, double frametime )
+// dimhotepus: double -> float as float always passed as argument.
+void CPlayerMove::CheckMovingGround( CBasePlayer *player, float frametime )
 {
 	VPROF( "CPlayerMove::CheckMovingGround()" );
